feat(a11y): Accessibility settings pane (Animation + Transparency + Character size)
